refactor: introduces a better project structure
